Create build_and_upload_wheels.yml and test_bamread.yml

+++ bamread/src/bamread.pyx
@@ -2,12 +2,12 @@
 
 # import sys
 
-import numpy as np
 import cython
-
+import numpy as np
 import pysam
 
-from libc.stdint cimport int32_t, uint32_t, uint64_t, int8_t, int16_t, uint16_t, int64_t
+from libc.stdint cimport (int8_t, int16_t, int32_t, int64_t, uint16_t,
+                          uint32_t, uint64_t)
 
 # from pysam.libcalignedsegment cimport AlignedSegment
